Move Fixture and FixtureManager to baaahs.fixtures.
