+++ src/commonMain/kotlin/baaahs/show/Show.kt
@@ -1,8 +1,8 @@
 package baaahs.show
 
-import baaahs.Fixture
 import baaahs.app.ui.Editable
 import baaahs.camelize
+import baaahs.fixtures.Fixture
 import baaahs.getBang
 import baaahs.plugin.Plugins
 import baaahs.show.ButtonGroupControl.Direction
